Stores - waitlist - persist
