+++ Apps/WebClient/src/NewClientApp/package.json
@@ -17,6 +17,7 @@
         "loglevel": "^1.8.1",
         "luxon": "^3.3.0",
         "pinia": "^2.0.23",
+        "pinia-plugin-persistedstate": "^3.1.0",
         "roboto-fontface": "*",
         "throttle-debounce": "^5.0.0",
         "vue": "^3.2.0",
