+++ Apps/WebClient/src/NewClientApp/src/stores/index.ts
@@ -1,4 +1,8 @@
 // Utilities
 import { createPinia } from "pinia";
+import piniaPluginPersistedState from "pinia-plugin-persistedstate";
 
-export default createPinia();
+const pinia = createPinia();
+pinia.use(piniaPluginPersistedState);
+
+export default pinia;
